# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface MealItem {
  name: string;
  servingDescription: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  /** AI confidence 0-1, only for AI-detected items */
  confidence?: number;
}

export type MealMealType = (typeof MealMealType)[keyof typeof MealMealType];

export const MealMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface Meal {
  id: number;
  mealType: MealMealType;
  loggedAt: string;
  imageUrl?: string | null;
  notes?: string | null;
  items: MealItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFat: number;
}

export type CreateMealBodyMealType =
  (typeof CreateMealBodyMealType)[keyof typeof CreateMealBodyMealType];

export const CreateMealBodyMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface CreateMealBody {
  mealType: CreateMealBodyMealType;
  items: MealItem[];
  imageUrl?: string | null;
  notes?: string | null;
  loggedAt?: string;
}

export type UpdateMealBodyMealType =
  (typeof UpdateMealBodyMealType)[keyof typeof UpdateMealBodyMealType];

export const UpdateMealBodyMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface UpdateMealBody {
  mealType?: UpdateMealBodyMealType;
  items?: MealItem[];
  notes?: string | null;
}

export interface AnalyzePhotoBody {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Image MIME type (image/jpeg, image/png, etc.) */
  mimeType: string;
}

export interface ClarificationQuestion {
  question: string;
  options: string[];
  /** Which detected item this question is about */
  itemIndex?: number;
}

export interface NutritionAnalysis {
  items: MealItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFat: number;
  /** Overall confidence 0-1 */
  confidence: number;
  needsClarification: boolean;
  clarificationQuestions: ClarificationQuestion[];
  analysisNotes?: string | null;
}

export type ClarifyMealBodyAnswersItem = {
  question: string;
  answer: string;
};

export interface ClarifyMealBody {
  originalAnalysis: NutritionAnalysis;
  imageBase64?: string | null;
  answers: ClarifyMealBodyAnswersItem[];
}

export interface BarcodeProduct {
  barcode: string;
  productName: string;
  brand?: string | null;
  servingSize: string;
  servingsPerContainer?: number | null;
  caloriesPerServing: number;
  proteinPerServing: number;
  carbsPerServing: number;
  fatPerServing: number;
  imageUrl?: string | null;
}

export interface Goals {
  dailyCalories: number;
  dailyProtein: number;
  dailyCarbs: number;
  dailyFat: number;
  updatedAt: string;
}

export interface UpdateGoalsBody {
  dailyCalories: number;
  dailyProtein: number;
  dailyCarbs: number;
  dailyFat: number;
}

export interface DailySummary {
  date: string;
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFat: number;
  goalCalories: number;
  goalProtein: number;
  goalCarbs: number;
  goalFat: number;
  meals: Meal[];
}

export type ListMealsParams = {
  /**
   * ISO date string YYYY-MM-DD, defaults to today
   */
  date?: string;
};

export type GetDailySummaryParams = {
  /**
   * ISO date string YYYY-MM-DD, defaults to today
   */
  date?: string;
};