# lib/api-zod/src/generated/types/analyzePhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzePhotoBody {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Image MIME type (image/jpeg, image/png, etc.) */
  mimeType: string;
}