# lib/api-zod/src/generated/types/barcodeProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BarcodeProduct {
  barcode: string;
  productName: string;
  brand?: string | null;
  servingSize: string;
  servingsPerContainer?: number | null;
  caloriesPerServing: number;
  proteinPerServing: number;
  carbsPerServing: number;
  fatPerServing: number;
  imageUrl?: string | null;
}