# lib/api-zod/src/generated/types/createMealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMealBodyMealType } from "./createMealBodyMealType";
import type { MealItem } from "./mealItem";

export interface CreateMealBody {
  mealType: CreateMealBodyMealType;
  items: MealItem[];
  imageUrl?: string | null;
  notes?: string | null;
  loggedAt?: Date;
}