# lib/api-zod/src/generated/types/goals.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Goals {
  dailyCalories: number;
  dailyProtein: number;
  dailyCarbs: number;
  dailyFat: number;
  updatedAt: Date;
}