# lib/api-zod/src/generated/types/meal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MealItem } from "./mealItem";
import type { MealMealType } from "./mealMealType";

export interface Meal {
  id: number;
  mealType: MealMealType;
  loggedAt: Date;
  imageUrl?: string | null;
  notes?: string | null;
  items: MealItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFat: number;
}