# lib/api-zod/src/generated/types/mealItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface MealItem {
  name: string;
  servingDescription: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  /** AI confidence 0-1, only for AI-detected items */
  confidence?: number;
}