# lib/api-zod/src/generated/types/nutritionAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ClarificationQuestion } from "./clarificationQuestion";
import type { MealItem } from "./mealItem";

export interface NutritionAnalysis {
  items: MealItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFat: number;
  /** Overall confidence 0-1 */
  confidence: number;
  needsClarification: boolean;
  clarificationQuestions: ClarificationQuestion[];
  analysisNotes?: string | null;
}